test: use native letsrunit steps
